+++ e2e/__tests__/eachError.test.ts
@@ -1,0 +1,31 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import * as path from 'path';
+import {wrap} from 'jest-snapshot-serializer-raw';
+import {extractSummary} from '../Utils';
+import runJest from '../runJest';
+
+const dir = path.resolve(__dirname, '../each');
+
+test('shows error message when not enough arguments are supplied to tests', () => {
+  const result = runJest(dir, ['eachException.test.js']);
+  expect(result.exitCode).toBe(1);
+  const {rest} = extractSummary(result.stderr);
+  expect(wrap(rest)).toMatchSnapshot();
+});
+
+test('shows the correct errors in stderr when failing tests', () => {
+  const result = runJest(dir, ['failure.test.js']);
+  expect(result.exitCode).toBe(1);
+  const output = extractSummary(result.stderr)
+    .rest.split('\n')
+    .map(line => line.trimRight())
+    .join('\n');
+  expect(wrap(output)).toMatchSnapshot();
+});
+
